+++ src/GitHub_Updater/Base.php
@@ -118,6 +118,12 @@ class Base {
 		} else {
 			self::$installed_apis['gitea_api'] = false;
 		}
+		if ( file_exists( __DIR__ . '/API/Gist_API.php' ) ) {
+			self::$installed_apis['gist_api'] = true;
+			self::$git_servers['gist']        = 'Gist';
+		} else {
+			self::$installed_apis = false;
+		}
 		if ( file_exists( __DIR__ . '/API/Zipfile_API.php' ) ) {
 			self::$installed_apis['zipfile_api'] = true;
 			self::$git_servers['zipfile']        = 'Zipfile';
@@ -607,7 +613,19 @@ class Base {
 			$rollback = array_keys( $rollback );
 			usort( $rollback, 'version_compare' );
 			krsort( $rollback );
-			$rollback = array_slice( $rollback, 0, 1 );
+
+			/**
+			 * Filter to return the number of tagged releases (rollbacks) in branch switching.
+			 *
+			 * @since 9.6.0
+			 * @param int Number of rollbacks. Zero implies value not set.
+			 */
+			$num_rollbacks = absint( apply_filters( 'github_updater_number_rollbacks', 0 ) );
+
+			// Still only return last tag if using release assets.
+			$rollback = 0 === $num_rollbacks || $data['release_asset']
+				? array_slice( $rollback, 0, 1 )
+				: array_splice( $rollback, 0, $num_rollbacks, true );
 
 			foreach ( $rollback as $tag ) {
 				printf(