+++ src/GitHub_Updater/Install.php
@@ -226,6 +226,13 @@ class Install {
 				self::$install = Singleton::get_instance( 'API\Zipfile_API', $this )->remote_install( $headers, self::$install );
 			}
 
+			/**
+			 * Install from Gist.
+			 */
+			if ( 'gist' === self::$install['github_updater_api'] ) {
+				self::$install = Singleton::get_instance( 'API\Gist_API', $this )->remote_install( $headers, self::$install );
+			}
+
 			if ( isset( self::$install['options'] ) ) {
 				$this->save_options_on_install( self::$install['options'] );
 			}
@@ -319,15 +326,6 @@ class Install {
 				? new CLI_Plugin_Installer_Skin()
 				: new \Plugin_Installer_Skin( compact( 'type', 'url', 'nonce', 'plugin' ) );
 			$upgrader = new \Plugin_Upgrader( $skin );
-			add_filter(
-				'install_plugin_complete_actions',
-				[
-					$this,
-					'install_plugin_complete_actions',
-				],
-				10,
-				3
-			);
 		}
 
 		if ( 'theme' === $type ) {
@@ -502,21 +500,6 @@ class Install {
 			</select>
 		</label>
 		<?php
-	}
-
-	/**
-	 * Remove activation links after plugin installation as no method to get $plugin_file.
-	 *
-	 * @param array  $install_actions Array of plugin actions.
-	 * @param mixed  $api             Unused.
-	 * @param string $plugin_file     Plugin file/slug.
-	 *
-	 * @return mixed
-	 */
-	public function install_plugin_complete_actions( $install_actions, $api, $plugin_file ) {
-		unset( $install_actions['activate_plugin'], $install_actions['network_activate'] );
-
-		return $install_actions;
 	}
 
 	/**