+++ src/GitHub_Updater/Traits/GHU_Trait.php
@@ -41,7 +41,7 @@ trait GHU_Trait {
 	 * @return bool
 	 */
 	public static function is_doing_ajax() {
-		return defined( 'DOING_AJAX' ) && DOING_AJAX;
+		return defined( 'DOING_AJAX' ) && \DOING_AJAX;
 	}
 
 	/**
@@ -400,12 +400,14 @@ trait GHU_Trait {
 			'Bitbucket' => 'bitbucket_' . $type,
 			'GitLab'    => 'gitlab_' . $type,
 			'Gitea'     => 'gitea_' . $type,
+			'Gist'      => 'gist_' . $type,
 		];
 		$repo_base_uris = [
 			'GitHub'    => 'https://github.com/',
 			'Bitbucket' => 'https://bitbucket.org/',
 			'GitLab'    => 'https://gitlab.com/',
 			'Gitea'     => '',
+			'Gist'      => 'https://gist.github.com/',
 		];
 
 		if ( array_key_exists( $repo, $repo_types ) ) {
@@ -666,7 +668,7 @@ trait GHU_Trait {
 			return true;
 		}
 		$dir = opendir( $source );
-		mkdir( $destination );
+		@mkdir( $destination );
 		$source = untrailingslashit( $source );
 		// phpcs:ignore WordPress.CodeAnalysis.AssignmentInCondition.FoundInWhileCondition
 		while ( false !== ( $file = readdir( $dir ) ) ) {