+++ CHANGES.md
@@ -1,9 +1,15 @@
 #### [unreleased]
 
-#### 9.5.2 / 2020-05-08
+#### 9.6.0 / 2020-06-01
+* add WP-CLI branch switching
+* keep _Activate Plugin_ link on remote install
+* add `class Gist_API` to install/update GitHub Gists, themes will use hash as slug
+* add filter `github_updater_number_rollbacks` to set the number of tagged releases (rollbacks) available in branch switching
+
+#### 9.5.2 / 2020-05-09
 * no need for using release asset with GHU
 
-#### 9.5.1 / 2020-05-08
+#### 9.5.1 / 2020-05-09
 * test `Readme_Parser::__construct()` `file_put_contents()` with additional test for success, hopefully squashes [#704](https://github.com/afragen/github-updater/issues/704) once and for all, actual fix is to set constant `WP_TEMP_DIR` as appropriate
 * prevent error if no credentials are set
 * un-screwup Bitbucket Server, sorry @allrite, fixes [#872](https://github.com/afragen/github-updater/issues/872)